+++ SarGoldACC.Core/Services/Interfaces/IGroupService.cs
@@ -5,9 +5,9 @@ namespace SarGoldACC.Core.Services.Interfaces;
 
 public interface IGroupService
 {
-    Task<GroupDto> GetByIdAsync(int id);
+    Task<GroupDto> GetByIdAsync(long id);
     Task<List<GroupDto>> GetAllAsync();
     Task AddAsync(GroupCreateDto createGroupDto);
     Task UpdateAsync(GroupDto groupDto);
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
 }
