+++ SarGoldACC.Core/Services/Interfaces/IPermissionService.cs
@@ -4,6 +4,6 @@ namespace SarGoldACC.Core.Services.Interfaces;
 
 public interface IPermissionService
 {
-    Task<PermissionDto> GetByIdAsync(int id);
+    Task<PermissionDto> GetByIdAsync(long id);
     Task<List<PermissionDto>> GetAllAsync();
 }
